added experiment class, and removed redundancies.
Added a csaodfExperiment class to handle all the interpolation, volume
management and so on, and included a runExperiment.m test script to
demonstrate how it works. Also renamed the SHmat class to
sphericalHarmonicsMatrix, so that it is more descriptive. Removed the
testDataSNR30.nii.gz as teh .nii file itself exists in the repo.